requirements.txt
django
djangorestframework
pillow
celery
redis
django-celery-results